Alterando instalação do gazebo e colocando if jammy

+++ debian-robotico.sh
@@ -3,7 +3,7 @@
 clear
 echo "####################################################################################"
 echo "##										##"
-echo "##                    Bem vindo a instalação do Ubuntu Robótico                   ##"
+echo "##                    Bem vindo a instalação do Debian Robótico                   ##"
 echo "##										##"
 echo "####################################################################################"
 echo ""
@@ -24,12 +24,12 @@ echo "##           Removendo programas pré-instalados CASO seja Lubuntu 22.04  
 echo "##										##"
 echo "####################################################################################"
 echo ""
-echo ""
-echo ""
-sudo apt autoremove qtpass screengrab 2048-qt quassel transmission-qt transmission-common skanlite picom noblenote compton info -y
-
+if [ $(lsb_release -cs) == 'jammy' ]; then
+  sudo apt autoremove qtpass screengrab 2048-qt quassel transmission-qt transmission-common skanlite picom noblenote compton info -y
+fi
 
 set -e #A partir de agora qualquer comando que retorar erro irá encerrar a execução do script
+
 
 
 clear
@@ -38,8 +38,10 @@ echo "##										##"
 echo "##                       Atualizando Sistema Operacional				##"
 echo "##										##"
 echo "####################################################################################"
+echo ""
 sudo apt update
 sudo apt upgrade -y
+
 
 
 clear
@@ -48,7 +50,28 @@ echo "##										##"
 echo "##                        Instalar Programas Básicos				##"
 echo "##										##"
 echo "####################################################################################"
-sudo apt install git chromium-browser deepin-icon-theme qt5-gtk2-platformtheme gnome-screenshot gedit gparted gimp inkscape vlc unrar curl geany geany-plugins -y
+echo ""
+sudo apt install git chromium-browser arc-theme deepin-icon-theme qt5-gtk2-platformtheme gnome-screenshot gedit gparted gimp inkscape vlc unrar curl geany geany-plugins -y
+
+
+
+clear
+echo "####################################################################################"
+echo "##										##"
+echo "##                            Instalando Spotify					##"
+echo "##										##"
+echo "####################################################################################"
+echo ""
+echo "---> Executando passo a passo fornecido pelo desenvolvedor <---"
+echo ""
+curl -sS https://download.spotify.com/debian/pubkey_5E3C45D7B312C643.gpg | sudo apt-key add - 
+echo "deb http://repository.spotify.com stable non-free" | sudo tee /etc/apt/sources.list.d/spotify.list
+sudo apt update
+echo "---> Baixando dependência faltante para Ubuntu 22.04 <---"
+wget http://archive.ubuntu.com/ubuntu/pool/main/o/openssl/libssl1.1_1.1.1l-1ubuntu1_amd64.deb
+sudo apt install spotify-client ./libssl1.1_1.1.1l-1ubuntu1_amd64.deb -y
+rm libssl1.1_1.1.1l-1ubuntu1_amd64.deb
+
 
 
 clear
@@ -72,53 +95,52 @@ sudo apt install code # or code-insiders
 
 
 clear
-echo "####################################################################################"
-echo "##										##"
-echo "##                            Instalando Spotify					##"
-echo "##										##"
-echo "####################################################################################"
-echo ""
-echo "---> Executando passo a passo fornecido pelo desenvolvedor <---"
-echo ""
-curl -sS https://download.spotify.com/debian/pubkey_5E3C45D7B312C643.gpg | sudo apt-key add - 
-echo "deb http://repository.spotify.com stable non-free" | sudo tee /etc/apt/sources.list.d/spotify.list
-sudo apt update
-echo "---> Baixando dependência faltante para Ubuntu 22.04 <---"
-wget http://archive.ubuntu.com/ubuntu/pool/main/o/openssl/libssl1.1_1.1.1l-1ubuntu1_amd64.deb
-sudo apt install spotify-client ./libssl1.1_1.1.1l-1ubuntu1_amd64.deb -y
-rm libssl1.1_1.1.1l-1ubuntu1_amd64.deb
+if [ $(lsb_release -cs) == 'jammy' ]; then
+  echo "####################################################################################"
+  echo "##										##"
+  echo "##                               Instalando ROS					##"
+  echo "##										##"
+  echo "####################################################################################"
+  echo ""
+  echo "---> Executando passo a passo fornecido pelo desenvolvedor <---"
+  echo ""
+  sudo apt install gnupg lsb-release
+  sudo curl -sSL https://raw.githubusercontent.com/ros/rosdistro/master/ros.key -o /usr/share/keyrings/ros-archive-keyring.gpg
+  echo "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/ros-archive-keyring.gpg] http://packages.ros.org/ros2/ubuntu $(source /etc/os-release && echo $UBUNTU_CODENAME) main" | sudo tee /etc/apt/sources.list.d/ros2.list > /dev/null
+  sudo apt update
+  sudo apt install ros-humble-desktop -y
+  source /opt/ros/humble/setup.bash
+fi
+
+
+
+#clear
+#echo "####################################################################################"
+#echo "##										##"
+#echo "##                             Instalando Gazebo					##"
+#echo "##										##"
+#echo "####################################################################################"
+#echo ""
+#echo "---> Executando passo a passo fornecido pelo desenvolvedor <---"
+#echo ""
+#sudo apt-get install lsb-release wget gnupg
+#sudo wget https://packages.osrfoundation.org/gazebo.gpg -O /usr/share/keyrings/pkgs-osrf-archive-keyring.gpg
+#echo "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/pkgs-osrf-archive-keyring.gpg] http://packages.osrfoundation.org/gazebo/ubuntu-stable $(lsb_release -cs) main" | sudo tee /etc/apt/sources.list.d/gazebo-stable.list > /dev/null
+#sudo apt-get update
+#sudo apt-get install ignition-fortress -y
+
 
 
 clear
 echo "####################################################################################"
 echo "##										##"
-echo "##                               Instalando ROS					##"
+echo "##                             Instalando Gazebo Classic			##"
 echo "##										##"
 echo "####################################################################################"
 echo ""
-echo "---> Executando passo a passo fornecido pelo desenvolvedor <---"
-echo ""
-sudo apt install gnupg lsb-release
-sudo curl -sSL https://raw.githubusercontent.com/ros/rosdistro/master/ros.key -o /usr/share/keyrings/ros-archive-keyring.gpg
-echo "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/ros-archive-keyring.gpg] http://packages.ros.org/ros2/ubuntu $(source /etc/os-release && echo $UBUNTU_CODENAME) main" | sudo tee /etc/apt/sources.list.d/ros2.list > /dev/null
-sudo apt update
-sudo apt install ros-humble-desktop -y
-source /opt/ros/humble/setup.bash
+sudo apt-get install gazebo -y
 
-clear
-echo "####################################################################################"
-echo "##										##"
-echo "##                             Instalando Gazebo					##"
-echo "##										##"
-echo "####################################################################################"
-echo ""
-echo "---> Executando passo a passo fornecido pelo desenvolvedor <---"
-echo ""
-sudo apt-get install lsb-release wget gnupg
-sudo wget https://packages.osrfoundation.org/gazebo.gpg -O /usr/share/keyrings/pkgs-osrf-archive-keyring.gpg
-echo "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/pkgs-osrf-archive-keyring.gpg] http://packages.osrfoundation.org/gazebo/ubuntu-stable $(lsb_release -cs) main" | sudo tee /etc/apt/sources.list.d/gazebo-stable.list > /dev/null
-sudo apt-get update
-sudo apt-get install ignition-fortress -y
+
 
 clear
 echo "####################################################################################"
@@ -134,6 +156,8 @@ echo ""
 echo "---> Iniciando instalação <---"
 echo ""
 sudo apt install ./webots_2022a_amd64.deb -y
+
+
 
 clear
 echo "####################################################################################"
